# .stryker-tmp/sandbox-nIGnBR/backend/docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Datacendia API
  description: |
    Enterprise Decision Intelligence Platform API
    
    ## Authentication
    All endpoints except `/health` and `/api/v1/auth/*` require a valid JWT token.
    Include the token in the `Authorization` header: `Bearer <token>`
    
    ## Rate Limiting
    - Standard tier: 100 requests/minute
    - Enterprise tier: 1000 requests/minute
    - Rate limit headers: `X-RateLimit-Limit`, `X-RateLimit-Remaining`, `X-RateLimit-Reset`
    
    ## Error Responses
    All errors follow a consistent format with `success: false` and an `error` object.
  version: 1.0.0
  contact:
    name: Datacendia Support
    email: support@datacendia.com
    url: https://datacendia.com/support
  license:
    name: Proprietary
    url: https://datacendia.com/license
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.datacendia.com
    description: Production server
...
paths:
...
  /api/v1/auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
              required: [refreshToken]
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/v1/auth/login
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: Authentication required
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
          required: [code, message]
      required: [success, error]
...
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            token:
              type: string
            refreshToken:
              type: string
            user:
              $ref: '#/components/schemas/User'
            expiresIn:
              type: integer
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        department:
          type: string
        createdAt:
          type: string
          format: date-time
    UserRole:
      type: string
      enum: [admin, analyst, operator, auditor, viewer]
...